начал перевод на шаблоны

+++ engine/basket.php
@@ -102,43 +102,47 @@ function setBasket(){
         }
 }
 function basketTable(){
-    $basket=unserialize($_COOKIE['basket_product']);
-    $mysqli = new mysqli(SQL_SERVER, SQL_USER, SQL_PASS, dbname, SQL_PORT);
-    foreach ($basket as $key => $value) {
-        $res = $mysqli->query("SELECT name FROM product WHERE id=" . $value['id']);
-        $res = $res->fetch_assoc();
-        $basket[$key]['name'] = $res['name'];
-    }
-    $str="<table><tr><td>Название</td><td>Цена</td></tr>";
-    $price=0;
-    foreach($basket as $key=>$value){
-        $price+=$value['price'];
-        $str.="<tr><td>".$value['name']."</td><td>".$value['price']."</td><td>
+    if (isset($_COOKIE['basket_product'])) {
+        $basket = unserialize($_COOKIE['basket_product']);
+        $mysqli = new mysqli(SQL_SERVER, SQL_USER, SQL_PASS, dbname, SQL_PORT);
+        foreach ($basket as $key => $value) {
+            $res = $mysqli->query("SELECT name FROM product WHERE id=" . $value['id']);
+            $res = $res->fetch_assoc();
+            $basket[$key]['name'] = $res['name'];
+        }
+        $str = "<table><tr><td>Название</td><td>Цена</td></tr>";
+        $price = 0;
+        foreach ($basket as $key => $value) {
+            $price += $value['price'];
+            $str .= "<tr><td>" . $value['name'] . "</td><td>" . $value['price'] . "</td><td>
         <button id='key' onclick='deleteBasketId($key)'>Удалить</button></td></tr>";
+        }
+        $str .= "<tr><td>ИТОГО</td><td>$price</td></tr></table>";
+        return $str;
     }
-    $str.="<tr><td>ИТОГО</td><td>$price</td></tr></table>";
-    return $str;
 }
 
 function deleteBasketId($key){
     $basket=unserialize($_COOKIE['basket_product']);
     setcookie('basket_product',"",time()-1,"/");
     unset($basket[$key]);
-    foreach($basket as $value){
-        $basket_new[]=$value;
-    }
-    unset ($basket);
-    setcookie('basket_product',serialize($basket_new),time()+3600,"/");
-    $mysqli = new mysqli(SQL_SERVER, SQL_USER, SQL_PASS, dbname, SQL_PORT);
-    $price=0;
-    $str="<table><tr><td>Название</td><td>Цена</td></tr>";
-    foreach ($basket_new as $key => $value) {
-        $res = $mysqli->query("SELECT name FROM product WHERE id=" . $value['id']);
-        $res = $res->fetch_assoc();
-        $price+=$value['price'];
-        $str.="<tr><td>".$res['name']."</td><td>".$value['price']."</td><td>
+    if(count($basket)>0) {
+        foreach ($basket as $value) {
+            $basket_new[] = $value;
+        }
+        unset ($basket);
+        setcookie('basket_product', serialize($basket_new), time() + 3600, "/");
+        $mysqli = new mysqli(SQL_SERVER, SQL_USER, SQL_PASS, dbname, SQL_PORT);
+        $price = 0;
+        $str = "<table><tr><td>Название</td><td>Цена</td></tr>";
+        foreach ($basket_new as $key => $value) {
+            $res = $mysqli->query("SELECT name FROM product WHERE id=" . $value['id']);
+            $res = $res->fetch_assoc();
+            $price += $value['price'];
+            $str .= "<tr><td>" . $res['name'] . "</td><td>" . $value['price'] . "</td><td>
         <button id='key' onclick='deleteBasketId($key)'>Удалить</button></td></tr>";
+        }
+        $str .= "<tr><td>ИТОГО</td><td>$price</td></tr></table>";
+        return $str;
     }
-    $str.="<tr><td>ИТОГО</td><td>$price</td></tr></table>";
-    return $str;
 }
